requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.1.12
django-cors-headers==3.5.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
drf-extra-fields==3.4.1
drf-social-auth==0.1.0
drf-social-oauth2==1.2.1
drf-writable-nested==0.7.0
ecdsa==0.18.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
jwcrypto==1.4.2
nodeenv==1.7.0
oauthlib==3.2.2
Pillow==9.4.0
pip==20.2.3
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12.3
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
setuptools==49.2.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.2.4
tzdata==2022.7
urllib3==1.26.14
wrapt==1.15.0

